pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ssp-toolkit"
version = "0.1.0"
description = "Scripts for managing the creation of an SSP"
readme = "README.md"
requires-python = ">=3.14"
license = "GPL-3.0 AND CC0-1.0"
keywords = ["ssp", "security", "compliance", "automation"]
authors = [
    {name = "Tom Camp", email = "tom.camp@civicactions.com"},
    {name = "Fen Lebalme", email = "fen@civicactions.com"},
    {name = "Alanna Blinn", email = "alanna.blinn@civicactions.com"},
]
dependencies = [
    "blinker>=1.9.0",
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "markdown>=3.10",
    "md-toc>=9.0.0",
    "pydantic>=2.12.5",
    "pypandoc>=1.16.2",
    "python-dotenv>=1.2.1",
    "rtyaml>=1.0.0",
    "ruamel-yaml>=0.19.1",
    "slugify>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "mypy>=1.5.1",
    "black>=24.11.0",
    "isort>=6.0.0",
    "flake8>=6.1.0",
    "pre-commit>=3.4.0",
]

[project.urls]
Source = "https://github.com/CivicActions/ssp-toolkit"

[project.scripts]
cli = "tools.cli:cli"
createfiles = "src.tools.createfiles.createfiles:main"
creatematrix = "src.tools.creatematrix.creatematrix:main"
createproject = "tools.create_project:create_project"
exportto = "src.tools.exportto.exportto:main"
makefamilies = "src.tools.makefamilies.makefamilies:main"
sop = "src.tools.sop.sop:main"
makessp = "src.tools.makessp.makessp:main"
getconfig = "src.tools.helpers.config:check_config"

[tool.mypy.overrides]
module = "yaml"
ignore_missing_imports = true
